Update to the latest python3-pcloudapi library
The prior versions of python3-pcloudapi allowed for redirects to be followed
to unprotected http resources, which could be a security problem under certain
circumstances for some versions of git-annex, see the git-annex advisory
https://git-annex.branchable.com/security/CVE-2018-10857_and_CVE-2018-10859/ .

The pcloud API is accessed using HTTPS which would make it hard for someone to
man-in-the-middle it and pcloud can be reasonably trusted not to be malicious.
Therefore there shouldn't be significant cause for concern.